my-project/src/backend/requirements.txt
# Python dependencies for backend services
supabase>=2.0.0
httpx>=0.25.0
python-dotenv>=1.0.0

